+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!--next-version-placeholder-->
 
+## [v3.8.1] - 2025-04-29
+### Fixed
+* Series.__getitem__
 ## [v3.8.0] - 2025-04-28
 ### Added
 * Novel sorting routine for n-dimensional DICOM datasets.